+++ tests/QueryTest.php
@@ -105,25 +105,25 @@ class QueryTest extends TestCase
     public function testFrom()
     {
         $this->assertEquals(
-            'users',
+            '`users`',
             $this->create->query()
                 ->from('users')
                 ->getFrom()
         );
         $this->assertEquals(
-            'users u',
+            '`users` u',
             $this->create->query()
                 ->from('users', 'u')
                 ->getFrom()
         );
         $this->assertEquals(
-            'users,groups',
+            '`users`,`groups`',
             $this->create->query()
                 ->from(['users', 'groups'])
                 ->getFrom()
         );
         $this->assertEquals(
-            'users u,groups g',
+            '`users` u,`groups` g',
             $this->create->query()
                 ->from(['u' => 'users', 'g' => 'groups'])
                 ->getFrom()
@@ -136,7 +136,7 @@ class QueryTest extends TestCase
     public function testInto()
     {
         $this->assertEquals(
-            'users',
+            '`users`',
             $this->create->query()
                 ->into('users')
                 ->getTable()
@@ -149,23 +149,23 @@ class QueryTest extends TestCase
     public function testJoin()
     {
         $this->assertSQL(
-            'users u INNER JOIN teams t ON u.team_id = t.id',
+            '`users` u INNER JOIN `teams` t ON u.team_id = t.id',
             $this->create->query()
                 ->from('users', 'u')
                 ->innerJoin('teams', 't', 'u.team_id = t.id')
                 ->getFrom()
         );
         $this->assertSQL(
-            'users u RIGHT JOIN teams t ON u.team_id = t.id',
+            '`users` u RIGHT JOIN `teams` t ON u.team_id = t.id',
             $this->create->query()
                 ->from('users', 'u')
                 ->rightJoin('teams', 't', 'u.team_id = t.id')
                 ->getFrom()
         );
         $this->assertSQL(
-            'users u
-             LEFT JOIN users_groups ug ON u.id = ug.user_id
-             LEFT JOIN groups g ON g.id = ug.group_id',
+            '`users` u
+             LEFT JOIN `users_groups` ug ON u.id = ug.user_id
+             LEFT JOIN `groups` g ON g.id = ug.group_id',
             $this->create->query()
                 ->from('users', 'u')
                 ->leftJoin('users_groups', 'ug', 'u.id = ug.user_id')
@@ -284,15 +284,15 @@ class QueryTest extends TestCase
 
         $this->assertInstanceOf(Query::class, $select);
         $this->assertSQL(
-            "SELECT 1 FROM dual",
+            "SELECT 1 FROM `dual`",
             $connection->select(1)->from('dual')->getQuery()
         );
         /** @noinspection SqlResolve */
         $this->assertSQL(
             "SELECT g.*, COUNT(1) as active_users
-             FROM users u
-             LEFT JOIN users_groups ug ON u.id = ug.user_id
-             LEFT JOIN groups g ON g.id = ug.group_id
+             FROM `users` u
+             LEFT JOIN `users_groups` ug ON u.id = ug.user_id
+             LEFT JOIN `groups` g ON g.id = ug.group_id
              WHERE users.active = '1'
              GROUP BY g.id
              HAVING COUNT(u.id) > 1
@@ -321,7 +321,7 @@ class QueryTest extends TestCase
         $connection
             ->expects($this->once())
             ->method('query')
-            ->with($this->sql("SELECT * FROM users WHERE id = '1'"))
+            ->with($this->sql("SELECT * FROM `users` WHERE id = '1'"))
             ->willReturn($result = new FetchedResult([['id' => 1, 'username' => 'john']]));
 
         $this->assertSame(
@@ -339,7 +339,7 @@ class QueryTest extends TestCase
         $connection
             ->expects($this->once())
             ->method('fetch')
-            ->with($this->sql("SELECT * FROM users WHERE id = '1'"))
+            ->with($this->sql("SELECT * FROM `users` WHERE id = '1'"))
             ->willReturn($result = new FetchedResult([['id' => 1, 'username' => 'john']]));
 
         $this->assertSame(
@@ -357,12 +357,12 @@ class QueryTest extends TestCase
         $insert     = $connection->insert('users');
 
         $this->assertInstanceOf(Query::class, $insert);
-        $this->assertSame('users', $insert->getTable());
+        $this->assertSame('`users`', $insert->getTable());
 
         $connection
             ->expects($this->once())
             ->method('execute')
-            ->with($this->sql("INSERT INTO users (`username`) VALUES ('sam')"))
+            ->with($this->sql("INSERT INTO `users` (`username`) VALUES ('sam')"))
             ->willReturn(1);
 
         $this->assertEquals(
@@ -380,23 +380,23 @@ class QueryTest extends TestCase
         $update     = $connection->update('users');
 
         $this->assertInstanceOf(Query::class, $update);
-        $this->assertSame('users', $update->getTable());
+        $this->assertSame('`users`', $update->getTable());
         /** @noinspection SqlResolve */
         $this->assertSQL(
-            "UPDATE users
+            "UPDATE `users`
              SET `active` = '0'",
             $update->set(['active' => 0])->getQuery()
         );
         /** @noinspection SqlResolve */
         $this->assertSQL(
-            "UPDATE users
+            "UPDATE `users`
              SET `active` = '0'
              WHERE email LIKE '%@example.com'",
             $update->where('email LIKE ?', '%@example.com')->getQuery()
         );
         /** @noinspection SqlResolve */
         $this->assertSQL(
-            "UPDATE users
+            "UPDATE `users`
              SET `active` = '0'
              WHERE email LIKE '%@example.com'
              ORDER BY id
@@ -407,7 +407,7 @@ class QueryTest extends TestCase
         $connection
             ->expects($this->once())
             ->method('execute')
-            ->with($this->sql("UPDATE users SET `username`='sam' WHERE `id`='2'"))
+            ->with($this->sql("UPDATE `users` SET `username`='sam' WHERE `id`='2'"))
             ->willReturn(1);
 
         $this->assertEquals(
@@ -425,24 +425,24 @@ class QueryTest extends TestCase
         $delete     = $connection->delete('users');
 
         $this->assertInstanceOf(Query::class, $delete);
-        $this->assertSame('users', $delete->getTable());
+        $this->assertSame('`users`', $delete->getTable());
         $this->assertSQL(
-            'DELETE FROM users',
+            'DELETE FROM `users`',
             $delete->getQuery()
         );
         $this->assertSQL(
-            "DELETE FROM users WHERE id='3'",
+            "DELETE FROM `users` WHERE id='3'",
             $delete->where('id = ?', 3)->getQuery()
         );
         $this->assertSQL(
-            "DELETE FROM users WHERE id='3' LIMIT 1",
+            "DELETE FROM `users` WHERE id='3' LIMIT 1",
             $delete->limit(1)->getQuery()
         );
 
         $connection
             ->expects($this->once())
             ->method('execute')
-            ->with($this->sql("DELETE FROM users WHERE `id`='3'"))
+            ->with($this->sql("DELETE FROM `users` WHERE `id`='3'"))
             ->willReturn(1);
 
         $this->assertEquals(
@@ -459,14 +459,14 @@ class QueryTest extends TestCase
         $connection = $this->create->connection();
 
         $this->assertSQL(
-            "SELECT 1 FROM dual",
+            "SELECT 1 FROM `dual`",
             $connection
                 ->select(1)
                 ->from('dual')
                 ->getQuery()
         );
         $this->assertSQL(
-            "SELECT 1 FROM dual",
+            "SELECT 1 FROM `dual`",
             (string) $connection
                 ->select(1)
                 ->from('dual')