+++ changelog.md
@@ -4,6 +4,10 @@ All notable changes to Neat Database components will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Fixed
+- Table identifier quoting in queries built using the Query builder.
+
 ## [1.1.0]
 ### Added
 - Query builder fetch method.